remove previous validators
